UI tweaks and correctness

+++ app/src/main/java/com/example/jared/smart_bandage_android/SmartBandage.java
@@ -128,7 +128,7 @@ public class SmartBandage implements Serializable{
     }
 
 
-    public static int parseSysTime(byte[] data){
+    public static long parseSysTime(byte[] data){
         return ReadingList.parse32BitLittleEndian(data, 0);
     }
